pyproject.toml
# pyproject.toml
# Production-ready Python package configuration using UV (astral-sh) standards
# For WatsonX VideoGenie - AI-Powered Video Generation Platform

[project]
name = "watsonx-videogenie"
version = "1.0.0"
description = "Enterprise-grade AI-powered video generation platform using IBM Watson X, converting text scripts into narrated avatar videos with GPU-accelerated rendering"
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "Apache-2.0" }
authors = [
    { name = "Ruslan Magana", email = "ruslan@ruslanmv.com" }
]
maintainers = [
    { name = "Ruslan Magana", email = "ruslan@ruslanmv.com" }
]
keywords = [
    "ai",
    "video-generation",
    "watson-x",
    "ibm-cloud",
    "fastapi",
    "microservices",
    "kubernetes",
    "gpu-rendering",
    "wav2lip",
    "text-to-video"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Video",
    "Framework :: FastAPI",
    "Typing :: Typed",
]

# Core runtime dependencies with pinned versions for reproducibility
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.29.0",
    "requests==2.32.3",
    "pydantic==2.7.3",
    "ibm-watsonx-ai>=0.4.0,<1.0.0",
    "kafka-python==2.0.2",
    "boto3==1.34.106",
    "python-multipart==0.0.9",
    "httpx==0.27.0",
    "aiofiles==23.2.1",
]

[project.optional-dependencies]
# Development dependencies - linting, formatting, testing
dev = [
    "black==24.4.2",
    "flake8==7.0.0",
    "isort==5.13.2",
    "mypy==1.10.0",
    "pytest==8.2.0",
    "pytest-asyncio==0.23.6",
    "pytest-cov==5.0.0",
    "pytest-mock==3.14.0",
    "pre-commit==3.7.1",
    "ruff==0.4.4",
]

# Testing dependencies
test = [
    "pytest==8.2.0",
    "pytest-asyncio==0.23.6",
    "pytest-cov==5.0.0",
    "pytest-mock==3.14.0",
    "httpx==0.27.0",
]

# Documentation dependencies
docs = [
    "mkdocs==1.6.0",
    "mkdocs-material==9.5.21",
    "mkdocstrings[python]==0.25.1",
]

# All optional dependencies combined
all = [
    "watsonx-videogenie[dev,test,docs]",
]

[project.urls]
Homepage = "https://ruslanmv.com"
Documentation = "https://github.com/ruslanmv/watsonx-videogenie/blob/main/README.md"
Repository = "https://github.com/ruslanmv/watsonx-videogenie"
Issues = "https://github.com/ruslanmv/watsonx-videogenie/issues"
Changelog = "https://github.com/ruslanmv/watsonx-videogenie/releases"

[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["services"]

[tool.uv]
# UV-specific configuration for package management
dev-dependencies = [
    "black==24.4.2",
    "flake8==7.0.0",
    "isort==5.13.2",
    "mypy==1.10.0",
    "pytest==8.2.0",
    "pytest-asyncio==0.23.6",
    "pytest-cov==5.0.0",
    "pytest-mock==3.14.0",
    "pre-commit==3.7.1",
    "ruff==0.4.4",
]

[tool.uv.sources]
# Ensure dependencies are resolved from PyPI
# UV will use the lock file for reproducible installs

# Black code formatter configuration
[tool.black]
line-length = 88
target-version = ["py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | services/avatar-service/app/Wav2Lip
)/
'''

# isort import sorting configuration
[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/Wav2Lip/*", "*/.venv/*", "*/venv/*"]

# Pytest configuration
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=services",
    "--cov=renderer",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "-vv",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
]
asyncio_mode = "auto"

# Coverage configuration
[tool.coverage.run]
source = ["services", "renderer"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/Wav2Lip/*",
    "*/.venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

# MyPy type checking configuration
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
show_error_context = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "kafka.*",
    "boto3.*",
    "botocore.*",
    "ibm_watsonx_ai.*",
]
ignore_missing_imports = true

# Ruff linter configuration (modern alternative to flake8)
[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate commented out code
    "PL",     # pylint
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "PLR0913", # too many arguments
    "PLR0912", # too many branches
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py
"tests/*" = ["ARG"]       # unused arguments in tests

[tool.ruff.isort]
known-first-party = ["services", "renderer"]

[tool.ruff.pydocstyle]
convention = "google"
